Refactor file structure: move inbox-related components to `inbox` directory and update imports accordingly
