+++ src/inbox/IterableInboxDataModel.ts
@@ -3,13 +3,13 @@ import {
   IterableHtmlInAppContent,
   IterableInAppDeleteSource,
   IterableInAppLocation,
-} from './IterableInAppClasses';
+} from '../IterableInAppClasses';
 
-import { Iterable } from './Iterable';
+import { Iterable } from '../Iterable';
 
-import type { InboxImpressionRowInfo } from './InboxImpressionRowInfo';
-import type { InboxRowViewModel } from './InboxRowViewModel';
-import IterableInAppMessage from './IterableInAppMessage';
+import type { InboxImpressionRowInfo } from '../InboxImpressionRowInfo';
+import type { InboxRowViewModel } from '../InboxRowViewModel';
+import IterableInAppMessage from '../IterableInAppMessage';
 
 const RNIterableAPI = NativeModules.RNIterableAPI;
