+++ src/inbox/IterableInboxEmptyState.tsx
@@ -1,9 +1,9 @@
 import { StyleSheet, Text, View } from 'react-native';
 
-import { type IterableInboxCustomizations } from './inbox/IterableInboxCustomizations';
+import { type IterableInboxCustomizations } from './IterableInboxCustomizations';
 
 // TODO: Comment
-type EmptyStateProps = {
+export interface IterableInboxEmptyStateProps {
   customizations: IterableInboxCustomizations;
   tabBarHeight: number;
   tabBarPadding: number;
@@ -11,7 +11,7 @@ type EmptyStateProps = {
   contentWidth: number;
   height: number;
   isPortrait: boolean;
-};
+}
 
 export const IterableInboxEmptyState = ({
   customizations,
@@ -20,7 +20,7 @@ export const IterableInboxEmptyState = ({
   navTitleHeight,
   height,
   isPortrait,
-}: EmptyStateProps) => {
+}: IterableInboxEmptyStateProps) => {
   const defaultTitle = 'No saved messages';
   const defaultBody = 'Check again later!';
