+++ src/inbox/index.ts
@@ -1,2 +1,4 @@
 export * from './IterableInbox';
 export * from './IterableInboxCustomizations';
+export * from './IterableInboxEmptyState';
+export * from './IterableInboxMessageCell';
